fix: 로딩 페이지 수정
