+++ src/App.js
@@ -14,8 +14,7 @@ import ErrorPage from './pages/ErrorPage';
 import MRankDetail from './components/rankingpage/M_RankDetail';
 import { Provider } from 'react-redux';
 import store from './store/store';
-import Loading2Page from './pages/Loading2';
-import LoadingPage from './pages/LoadingPage';
+
 function App() {
     return (
         <>
@@ -51,11 +50,6 @@ function App() {
                             element={<AboutUsPage />}
                         />
                         <Route path='/menu' element={<MenuPage />} />
-                        <Route
-                            path='/loading2'
-                            element={<Loading2Page />}
-                        />{' '}
-                        <Route path='/loading1' element={<LoadingPage />} />
                         <Route path='/*' element={<ErrorPage />} />
                     </Routes>
                 </BrowserRouter>
